Makes easy to install and execute command line relying on Node.JS.
An extension can now detect when any CommandLineConverterEngine is created to download and install NodeJS and NPM into the working directory. This will allow to make the "npm" binary available transparently and gives a control over the version to use. The extension relies on "frontend-plugin-core" maven dependency to install the archives.
 Fixes #215.