Completely reviews HTML parsing strategy.
The existing HTML parser was not reliable according to #186 and #182. The HtmlInspectorEngine was parsing the HTML document with regex, which did provide perfect results.
A draft of new HTML parser has been created on a separated branch (html-line-inspector) but is not finished yet. Until an out of the box solution is released, a plugable API has been introduced to allow the registration of HTML parsers relying of third party libraries. One plugin has been created as an extension (wuic-attoparser) that relies on AttoParser project. The library is more reliable than the internal parser and it fixes #186 and #182. This extension is now required to detect web assets in any HTML document that is parsed (directly by the HtmlInspectorEngine or through the HtmlParserFilter). If the extension is not added to the classpath, then the HtmlParserInspector does not inspects the HTML and logs a warning.